Moved postman utils method to dedicated class
